requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.8
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanize==4.9.0
identify==2.5.33
idna==3.6
kombu==5.3.4
nodeenv==1.8.0
platformdirs==4.1.0
pre-commit==3.6.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
ruff==0.1.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
vine==5.1.0
virtualenv==20.25.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
